Rename files to match CommonJS conventions
