+++ src/client/app/js/controllers/error.js
