+++ src/client/app/js/controllers/guide.js
