+++ src/client/app/js/directives/index.js
@@ -1,4 +1,4 @@
 module.exports = angular.module('app.directives', [
   '$strap.directives',
-  require('./issuesDirective').name
+  require('./issues').name
 ]);
