+++ src/client/app/js/services/github.js
